feat: implement validation framework with DataAnnotations and custom validators
