+++ src/MediatorLite/Configuration/MediatorOptions.cs
@@ -84,7 +84,16 @@ public sealed class MediatorOptions
                 $"Type {behaviorType.Name} must have exactly 2 generic type parameters.",
                 nameof(behaviorType));
         }
-
+        
+        var implementsPipelineBehavior = behaviorType.GetInterfaces()
+            .Any(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IPipelineBehavior<,>));
+        if (!implementsPipelineBehavior)
+        {
+            throw new ArgumentException(
+                $"Type '{behaviorType.Name}' must implement IPipelineBehavior<TRequest, TResponse>.",
+                nameof(behaviorType));
+        }
+        
         _behaviorTypes.Add(behaviorType);
         return this;
     }
