+++ src/MediatorLite/Validation/Models/ValidationError.cs
@@ -1,0 +1,12 @@
+namespace MediatorLite.Validation.Models;
+
+/// <summary>
+/// Represents a single validation error.
+/// </summary>
+/// <param name="PropertyName">The name of the property that failed validation.</param>
+/// <param name="ErrorMessage">The error message describing the validation failure.</param>
+/// <param name="AttemptedValue">The value that was attempted.</param>
+public sealed record ValidationError(
+    string PropertyName,
+    string ErrorMessage,
+    object? AttemptedValue = null);
